+++ casageo/coder.py
@@ -19,12 +19,11 @@ This module provides geocoding and search operations.
 """
 
 import argparse
-import contextlib
 import logging
 import os
 import statistics
 import sys
-from collections.abc import Generator, Iterable, Mapping, Sequence
+from collections.abc import Mapping, Sequence
 from typing import Any, Final, cast
 
 from geopandas import GeoDataFrame
@@ -36,6 +35,7 @@ from casageo.tools._util import (
     and_then,
     dict_to_point,
     getpoint,
+    list_first,
     point_xy,
     split_if_str,
     to_records,
@@ -88,18 +88,6 @@ def _coder_params(q: Mapping) -> dict:
     }
 
 
-def _split_navigation(items: Iterable[dict]) -> Generator[dict]:
-    for item in items:
-        if nav_positions := item.get("access", []):
-            for nav_pos in nav_positions:
-                yield item | {"access": nav_pos}
-        else:
-            single = item.copy()
-            with contextlib.suppress(KeyError):
-                del single["access"]
-            yield single
-
-
 class AddressResult(CasaGeoResult):
     """
     Represents the result of a singular ``address()`` query.
@@ -114,6 +102,7 @@ class AddressResult(CasaGeoResult):
         address_details: bool = False,
         coordinates: bool = False,
         match_quality: bool = False,
+        navigation_points: bool = False,
         error_info: bool = False,
     ) -> GeoDataFrame:
         """
@@ -128,6 +117,7 @@ class AddressResult(CasaGeoResult):
             address_details: Include additional address details in the result.
             coordinates: Include numeric coordinate columns in the result.
             match_quality: Include match quality scores in the result.
+            navigation_points: Include navigation points in the result.
             error_info: Include error information in the result.
 
         Returns:
@@ -138,9 +128,8 @@ class AddressResult(CasaGeoResult):
         if id_ is None:
             id_ = 1
 
-        data: list[dict] = []
-        for index, item in enumerate(_split_navigation(self._data.get("items", [{}]))):
-            data.append(row := {})
+        def item2row(item, subid: int, navid: int = 0) -> dict[str, Any]:
+            row = {}
 
             address = item.get("address", {})
             scoring = item.get("scoring", {})
@@ -148,11 +137,12 @@ class AddressResult(CasaGeoResult):
             if True:
                 # fmt: off
                 row["id"]               = id_
-                row["subid"]            = index
+                row["subid"]            = subid
+                row["navid"]            = navid
                 row["address"]          = item.get("title")
                 row["resulttype"]       = item.get("resultType")
                 row["position"]         = and_then(item.get("position"), dict_to_point)
-                row["navigation"]       = and_then(item.get("access"), dict_to_point)
+                row["navigation"]       = and_then(item.get("access"), list_first, dict_to_point)
                 row["distance"]         = item.get("distance")
                 row["relevance"]        = scoring.get("queryScore")
                 row["timestamp"]        = self._timestamp
@@ -214,6 +204,17 @@ class AddressResult(CasaGeoResult):
                 row["mq_ontologyname"]  = score.get("ontologyName")
                 # fmt: on
 
+            return row
+
+        data: list[dict] = []
+        for subid, item in enumerate(self._data.get("items", [{}])):
+            data.append(item2row(item, subid))
+
+            if navigation_points:
+                for navindex, navpos in enumerate(item.get("access", [])):
+                    navitem = item | {"position": navpos, "access": None}
+                    data.append(item2row(navitem, subid, navindex + 1))
+
         if not data:
             return GeoDataFrame()
 
@@ -243,6 +244,7 @@ class PoiResult(CasaGeoResult):
         address_details: bool = False,
         coordinates: bool = False,
         category_codes: bool = False,
+        navigation_points: bool = False,
         error_info: bool = False,
     ) -> GeoDataFrame:
         """
@@ -257,6 +259,7 @@ class PoiResult(CasaGeoResult):
             address_details: Include additional address details in the result.
             coordinates: Include numeric coordinate columns in the result.
             category_codes: Include category, chain and food type IDs in the result.
+            navigation_points: Include navigation points in the result.
             error_info: Include error information in the result.
 
         Returns:
@@ -267,20 +270,20 @@ class PoiResult(CasaGeoResult):
         if id_ is None:
             id_ = 1
 
-        data: list[dict] = []
-        for index, item in enumerate(_split_navigation(self._data.get("items", [{}]))):
-            data.append(row := {})
+        def item2row(item, subid: int, navid: int = 0) -> dict[str, Any]:
+            row = {}
 
             address = item.get("address", {})
 
             if True:
                 # fmt: off
                 row["id"]               = id_
-                row["subid"]            = index
+                row["subid"]            = subid
+                row["navid"]            = navid
                 row["title"]            = item.get("title")
                 row["resulttype"]       = item.get("resultType")
                 row["position"]         = and_then(item.get("position"), dict_to_point)
-                row["navigation"]       = and_then(item.get("access"), dict_to_point)
+                row["navigation"]       = and_then(item.get("access"), list_first, dict_to_point)
                 row["distance"]         = item.get("distance")
                 row["timestamp"]        = self._timestamp
                 # fmt: on
@@ -323,6 +326,17 @@ class PoiResult(CasaGeoResult):
                 row["here_chains"]      = and_then(item.get("chains"),     lambda cs: [c["id"] for c in cs])
                 row["here_foodtypes"]   = and_then(item.get("foodTypes"),  lambda cs: [c["id"] for c in cs])
                 # fmt: on
+
+            return row
+
+        data: list[dict[str, Any]] = []
+        for subid, item in enumerate(self._data.get("items", [{}])):
+            data.append(item2row(item, subid))
+
+            if navigation_points:
+                for navindex, navpos in enumerate(item.get("access", [])):
+                    navitem = item | {"position": navpos, "access": None}
+                    data.append(item2row(navitem, subid, navindex + 1))
 
         if not data:
             return GeoDataFrame()
@@ -388,6 +402,7 @@ def address(
     address_details: bool = False,
     coordinates: bool = False,
     match_quality: bool = False,
+    navigation_points: bool = False,
 ) -> GeoDataFrame:
     """
     Geocode addresses.
@@ -401,6 +416,7 @@ def address(
         address_details: Include additional address details in the result.
         coordinates: Include numeric coordinate columns in the result.
         match_quality: Include match quality scores in the result.
+        navigation_points: Include navigation points in the result.
 
     Returns:
         ~geopandas.GeoDataFrame: The list of results as an EPSG:4326
@@ -422,6 +438,7 @@ def address(
         address_details=address_details,
         coordinates=coordinates,
         match_quality=match_quality,
+        navigation_points=navigation_points,
     ).dataframe()
     return cast(GeoDataFrame, df)
 
@@ -434,6 +451,7 @@ def address_result(
     address_details: bool = False,
     coordinates: bool = False,
     match_quality: bool = False,
+    navigation_points: bool = False,
 ) -> MultiResult:
     """:meta private:"""
 
@@ -449,6 +467,7 @@ def address_result(
         "address_details": address_details,
         "coordinates": coordinates,
         "match_quality": match_quality,
+        "navigation_points": navigation_points,
     }
 
     json = client.request(
@@ -488,6 +507,7 @@ def poi(
     address_details: bool = False,
     coordinates: bool = False,
     category_codes: bool = False,
+    navigation_points: bool = False,
 ) -> GeoDataFrame:
     """
     Search for points of interest (POI) around given positions.
@@ -501,6 +521,7 @@ def poi(
         address_details: Include additional address details in the result.
         coordinates: Include numeric coordinate columns in the result.
         category_codes: Include HERE category, chain and food type identifiers in the result.
+        navigation_points: Include navigation points in the result.
 
     Returns:
         ~geopandas.GeoDataFrame: The list of results as an EPSG:4326
@@ -522,6 +543,7 @@ def poi(
         address_details=address_details,
         coordinates=coordinates,
         category_codes=category_codes,
+        navigation_points=navigation_points,
     ).dataframe()
     return cast(GeoDataFrame, df)
 
@@ -534,6 +556,7 @@ def poi_result(
     address_details: bool = False,
     coordinates: bool = False,
     category_codes: bool = False,
+    navigation_points: bool = False,
 ) -> MultiResult:
     """:meta private:"""
 
@@ -549,6 +572,7 @@ def poi_result(
         "address_details": address_details,
         "coordinates": coordinates,
         "category_codes": category_codes,
+        "navigation_points": navigation_points,
     }
 
     json = client.request(
@@ -670,6 +694,11 @@ def _main(args: Sequence[str] | None = None) -> None:
         action="store_true",
         help="include match quality information in the output",
     )
+    address_dfparams.add_argument(
+        "--with-navigation-points",
+        action="store_true",
+        help="include navigation points in the output",
+    )
 
     # POI Subcommand
 
@@ -704,6 +733,11 @@ def _main(args: Sequence[str] | None = None) -> None:
         "--with-category-codes",
         action="store_true",
         help="include HERE category codes in the output",
+    )
+    poi_dfparams.add_argument(
+        "--with-navigation-points",
+        action="store_true",
+        help="include navigation points in the output",
     )
 
     # Main Function Body